// src/server/api/models/crm/models/PageTemplateClientGroup.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { ClientGroup } from './ClientGroup';

export type PageTemplateClientGroup = {
    /**
     * The page number requested. A zero-based page index (0..N)
     */
    page?: number;
    /**
     * The size of the page to be returned
     */
    size?: number;
    /**
     * The total number of pages available
     */
    totalPages?: number;
    /**
     * The total number of records available
     */
    totalCount?: number;
    /**
     * The field used to sort the response
     */
    sort?: string;
    /**
     * The sort direction
     */
    order?: PageTemplateClientGroup.order;
    /**
     * The records returned
     */
    content?: Array<ClientGroup>;
};

export namespace PageTemplateClientGroup {

    /**
     * The sort direction
     */
    export enum order {
        ASC = 'asc',
        DESC = 'desc',
    }


}